requirements.txt
datasets
pandas
chromadb
langchain
streamlit
google-generativeai
huggingface
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=1.12.0
pypdf
pytesseract>=0.3.10
Pillow>=9.4.0
pdf2image>=1.16.3
uvicorn==0.22.0